pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ez-yaml-config"
description = "Typed config files"
requires-python = ">=3.13"
classifiers = [
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
]
version = "0.2.0"
dependencies = [
    "pyyaml>=6.0.2",
    "pydantic>=2.11.4",
    "google-cloud-secret-manager>=2.23.3",
    "google-crc32c>=1.7.1",
]
[project.optional-dependencies]
dev = [
    "pre-commit>=4.2.0",
    "pylint>=3.3.6",
    "ruff==0.12.0",
    "uv==0.7.13",
]

[project.urls]
Homepage = "https://github.com/quarterzip/ez-yaml-config"
Issues = "https://github.com/quarterzip/ez-yaml-config/issues"

[tool.ruff]
line-length = 120
target-version = "py312"
[tool.ruff.lint]
# Allow applying these unsafe fixes without the `--unsafe-fixes` flag
extend-safe-fixes = ["F401"]
extend-select = [
    "I",   # isort
    "UP",  # pyupgrade
    "TID", # flake8-tidy-imports
]
